Extract duplicate clap helpers into clap-utils
