+++ archiver/Cargo.toml
@@ -10,6 +10,7 @@ homepage = "https://solana.com/"
 [dependencies]
 clap = "2.33.0"
 console = "0.9.1"
+solana-clap-utils = { path = "../clap-utils", version = "0.21.0" }
 solana-core = { path = "../core", version = "0.21.0" }
 solana-logger = { path = "../logger", version = "0.21.0" }
 solana-metrics = { path = "../metrics", version = "0.21.0" }
