+++ clap-utils/src/input_validators.rs
@@ -34,3 +34,17 @@ pub fn is_url(string: String) -> Result<(), String> {
         Err(err) => Err(format!("{:?}", err)),
     }
 }
+
+pub fn is_semver(semver: &str) -> Result<(), String> {
+    match semver::Version::parse(&semver) {
+        Ok(_) => Ok(()),
+        Err(err) => Err(format!("{:?}", err)),
+    }
+}
+
+pub fn is_release_channel(channel: &str) -> Result<(), String> {
+    match channel {
+        "edge" | "beta" | "stable" => Ok(()),
+        _ => Err(format!("Invalid release channel {}", channel)),
+    }
+}
