+++ cli/Cargo.toml
@@ -27,6 +27,7 @@ serde_derive = "1.0.102"
 serde_json = "1.0.41"
 serde_yaml = "0.8.11"
 solana-budget-api = { path = "../programs/budget_api", version = "0.21.0" }
+solana-clap-utils = { path = "../clap-utils", version = "0.21.0" }
 solana-client = { path = "../client", version = "0.21.0" }
 solana-config-api = { path = "../programs/config_api", version = "0.21.0" }
 solana-drone = { path = "../drone", version = "0.21.0" }
