+++ cli/src/cli.rs
@@ -1,13 +1,14 @@
 use crate::{
-    cluster_query::*, display::println_name_value, input_parsers::*, input_validators::*, stake::*,
-    storage::*, validator_info::*, vote::*,
+    cluster_query::*, display::println_name_value, stake::*, storage::*, validator_info::*, vote::*,
 };
 use chrono::prelude::*;
 use clap::{App, AppSettings, Arg, ArgMatches, SubCommand};
 use log::*;
 use num_traits::FromPrimitive;
 use serde_json::{self, json, Value};
+
 use solana_budget_api::budget_instruction::{self, BudgetError};
+use solana_clap_utils::{input_parsers::*, input_validators::*};
 use solana_client::{client_error::ClientError, rpc_client::RpcClient};
 #[cfg(not(test))]
 use solana_drone::drone::request_airdrop_transaction;
@@ -32,6 +33,7 @@ use solana_sdk::{
 use solana_stake_api::stake_state::{Lockup, StakeAuthorize};
 use solana_storage_api::storage_instruction::StorageAccountType;
 use solana_vote_api::vote_state::VoteAuthorize;
+
 use std::{
     fs::File,
     io::{Read, Write},
