+++ cli/src/lib.rs
@@ -5,8 +5,6 @@ pub mod cli;
 pub mod cluster_query;
 pub mod config;
 pub mod display;
-pub mod input_parsers;
-pub mod input_validators;
 pub mod stake;
 pub mod storage;
 pub mod validator_info;
