+++ install/Cargo.toml
@@ -23,11 +23,11 @@ lazy_static = "1.4.0"
 log = "0.4.8"
 nix = "0.15.0"
 reqwest = { version = "0.9.22", default-features = false, features = ["rustls-tls"] }
-semver = "0.9.0"
 serde = "1.0.102"
 serde_derive = "1.0.102"
 serde_yaml = "0.8.11"
 sha2 = "0.8.0"
+solana-clap-utils = { path = "../clap-utils", version = "0.21.0" }
 solana-client = { path = "../client", version = "0.21.0" }
 solana-config-api = { path = "../programs/config_api", version = "0.21.0" }
 solana-logger = { path = "../logger", version = "0.21.0" }
