+++ validator/Cargo.toml
@@ -17,6 +17,7 @@ log = "0.4.8"
 indicatif = "0.13.0"
 reqwest = { version = "0.9.22", default-features = false }
 serde_json = "1.0.41"
+solana-clap-utils = { path = "../clap-utils", version = "0.21.0" }
 solana-client = { path = "../client", version = "0.21.0" }
 solana-core = { path = "../core", version = "0.21.0" }
 solana-drone = { path = "../drone", version = "0.21.0" }
